fix success dialog disappearing after refresh finishes, fix selecting newest entry after creation, refactor
